1. Updated GH Pages badge and deployment section in README. 2. Added 'Beat to Rank Up' column to ExplorePlayers.
