+++ src/components/gear/BallsTable.js
@@ -1,13 +1,7 @@
 import React from "react";
 
-import {
-  generateBrandOptions,
-  generateColorOptions,
-  generateRarityOptions,
-  SelectColumnFilter,
-  TextColumnFilter
-} from "src/components/common/Table";
-import { BaseTable } from "src/components/gear/BaseTable";
+import { SelectColumnFilter, TextColumnFilter } from "src/components/common/Table";
+import { BaseTable, generateBrandOptions, generateColorOptions, generateRarityOptions } from "src/components/gear/BaseTable";
 
 export function BallsTable({ theme, data }) {
   const columns = React.useMemo(
