+++ src/components/gear/EyewearTable.js
@@ -1,13 +1,7 @@
 import React from "react";
 
-import {
-  generateBrandOptions,
-  generateColorOptions,
-  generateRarityOptions,
-  generateStyleOptions,
-  SelectColumnFilter
-} from "src/components/common/Table";
-import { BaseTable } from "src/components/gear/BaseTable";
+import { SelectColumnFilter } from "src/components/common/Table";
+import { BaseTable, generateBrandOptions, generateColorOptions, generateRarityOptions, generateStyleOptions } from "src/components/gear/BaseTable";
 
 export function EyewearTable({ theme, data }) {
   const columns = React.useMemo(
